Add missing app module build script and fix two build-blocking source errors
The project could not produce an APK: app/build.gradle.kts was committed as
an empty file, so the :app module had no Android plugin, no SDK levels and no
dependencies at all.

- app/build.gradle.kts: author the module build script matching the stack
  documented in README (compileSdk/targetSdk 36, minSdk 26, Java 17, Compose
  BOM 2025.09.01, CameraX 1.6.2, KomoUI 0.4.0 as io.github.derangga:komoui,
  material-icons-extended for the Outlined icons the UI uses).
- app/proguard-rules.pro: add the file the release build type references.
- FilterProcessor.kt: 0x9E3779B9 exceeds Int.MAX_VALUE so Kotlin inferred
  Long, which does not match deterministicNoise(x, y, seed: Int). Narrow it
  with .toInt() to keep the intended constant's bit pattern.
- MainActivity.kt: KomoTheme lives in com.komoui.themes, not com.komoui.
- settings.gradle.kts: drop the develocity {} block. No settings plugin
  provides that extension here, so configuration failed before any task ran.
- gradlew: restore the executable bit.

Verified with :app:assembleDebug and :app:assembleRelease. The debug APK
verifies under APK Signature Scheme v2 and exposes the expected launcher
activity.

+++ app/src/main/java/com/janoski/retroframe/FilterProcessor.kt
@@ -22,7 +22,7 @@ object FilterProcessor {
         val grain = (preset.grain * intensity).coerceIn(0f, 1f)
         val vignette = (preset.vignette * intensity).coerceIn(0f, 1f)
 
-        val seed = 0x9E3779B9
+        val seed = 0x9E3779B9.toInt()
         val centerX = (w - 1) * 0.5f
         val centerY = (h - 1) * 0.5f
         val maxDist = sqrt(centerX * centerX + centerY * centerY)
